[Dagger refactor] Split ComponentHjarGenerator from ComponentHjarProcessingStep.
This is in preparation to move component generation logic into its own sub-package.

Instead of ComponentHjarProcessingStep using the ComponentHjarGenerator directly (note: it was previously called EmptyComponentGenerator), it now injects SourceFileGenerator<ComponentDescriptor>, similar to what is done for ComponentProcessingStep and ComponentGenerator via SourceFileGenerator<BindingGraph>.

In a follow-up CL, I'll move all of the component generator logic into a subpackage.

RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=321894788
