+++ java/dagger/internal/codegen/SourceFileGeneratorsModule.java
@@ -16,10 +16,12 @@
 
 package dagger.internal.codegen;
 
+import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import dagger.internal.codegen.base.SourceFileGenerator;
 import dagger.internal.codegen.binding.BindingGraph;
+import dagger.internal.codegen.binding.ComponentDescriptor;
 import dagger.internal.codegen.binding.MembersInjectionBinding;
 import dagger.internal.codegen.binding.ProductionBinding;
 import dagger.internal.codegen.binding.ProvisionBinding;
@@ -36,32 +38,40 @@ import javax.lang.model.element.TypeElement;
 abstract class SourceFileGeneratorsModule {
 
   @Provides
-  static SourceFileGenerator<ProvisionBinding> provisionBindingGenerator(
+  static SourceFileGenerator<ProvisionBinding> factoryGenerator(
       FactoryGenerator generator, CompilerOptions compilerOptions) {
     return hjarWrapper(generator, compilerOptions);
   }
 
   @Provides
-  static SourceFileGenerator<ProductionBinding> productionBindingGenerator(
+  static SourceFileGenerator<ProductionBinding> producerFactoryGenerator(
       ProducerFactoryGenerator generator, CompilerOptions compilerOptions) {
     return hjarWrapper(generator, compilerOptions);
   }
 
   @Provides
-  static SourceFileGenerator<MembersInjectionBinding> membersInjectionBindingGenerator(
+  static SourceFileGenerator<MembersInjectionBinding> membersInjectorGenerator(
       MembersInjectorGenerator generator, CompilerOptions compilerOptions) {
     return hjarWrapper(generator, compilerOptions);
   }
 
-  @Provides
-  static SourceFileGenerator<BindingGraph> bindingGraphGenerator(
-      ComponentGenerator generator, CompilerOptions compilerOptions) {
-    return hjarWrapper(generator, compilerOptions);
-  }
+  @Binds
+  abstract SourceFileGenerator<BindingGraph> componentGenerator(ComponentGenerator generator);
+
+  // The HjarSourceFileGenerator wrapper first generates the entire TypeSpec before stripping out
+  // things that aren't needed for the hjar. However, this can be really expensive for the component
+  // because it is usually the most expensive file to generate, and most of its content is not
+  // needed in the hjar. Thus, instead of wrapping the ComponentGenerator in HjarSourceFileGenerator
+  // we provide a completely separate processing step, ComponentHjarProcessingStep, and generator,
+  // ComponentHjarGenerator, for when generating hjars for components, which can avoid generating
+  // the parts of the component that would have been stripped out by the HjarSourceFileGenerator.
+  @Binds
+  abstract SourceFileGenerator<ComponentDescriptor> componentHjarGenerator(
+      ComponentHjarGenerator hjarGenerator);
 
   @Provides
   @ModuleGenerator
-  static SourceFileGenerator<TypeElement> moduleProxyGenerator(
+  static SourceFileGenerator<TypeElement> moduleConstructorProxyGenerator(
       ModuleConstructorProxyGenerator generator, CompilerOptions compilerOptions) {
     return hjarWrapper(generator, compilerOptions);
   }
